+++ src/test/kotlin/org/jetbrains/intellij/IntelliJPluginSpecBase.kt
@@ -106,7 +106,7 @@ abstract class IntelliJPluginSpecBase {
             .withPluginClasspath()
             .withDebug(debugEnabled)
             .withTestKitDir(File(gradleHome))
-            .withArguments(*tasks, "--stacktrace")//, "-Dorg.gradle.debug=true")
+            .withArguments(*tasks, "--stacktrace", "--configuration-cache")//, "-Dorg.gradle.debug=true")
 
     fun tasks(groupName: String): List<String> = build(ProjectInternal.TASKS_TASK).output.lines().run {
         val start = indexOfFirst { it.equals("$groupName tasks", ignoreCase = true) } + 2